Initial stab at having Utils support multiple loggers
If OpenLog() is called with an already defined logger, then Utils will
switch to that logger object.  If the given name is not defined, then
it will create a new logging object and switch to that logging object.

CAVEAT: If the currently defined logging object is
Log::Fine::Handle::File based, then when the logger is switched, the
previous File logger goes out of scope and closes down.  One
workaround is to set {autoclose} to 1 which forces recreation of the
internal file handle.  This will need to be addressed before release,
probably by having Utils keep track of its own logging objects or else
storing a ref to Log::Fine internally.

Other changes:

 - Add a new function, ListLoggers(), which returns a list of
   stored loggers.

 - Have File.pm and children check to make sure the internally stored
   filehandle is open via fileno.