Use the mailx(1) command for grabbing headers under solaris
Use the mailx(1) command for grabbing headers under solaris as the
stock mail(1) command does not support the '-H' option.  Tested under
OpenSolaris.

NOTE: A cursory check of IBM AIX[1] shows that their implementation of
mail /does/ support the '-H' option, so no need to check for it
there.  The mileage on other SysV-ish systems may vary.

Footnotes:
[1]  Not to say that Log::Fine officially supports IBM AIX.  The
author does not have access to an AIX box with a recent version of
perl on it.