Use File::Temp::mkstemp() to create temporary file
Use File::Temp::mkstemp() to create temporary file under Windows.
Also, instead of checking to see if a file is a plain file, just check
to see if it exists, which was causing a test to fail.  Finally, add
note at top of document to remind myself why these tests exist ;)