Properly call mail(1) when using Debian GNU/Linux
The mail(1) program under Debian GNU/Linux, like the OpenBSD version,
lacks support for the '-H' option, so utilize the same OpenBSD hack to
count the number of messages in the mail queue.

Note that, interestingly enough, that this does not appear to be the
case under the version of mail(1) provided by Ubuntu Linux 12.04,
which is Debian-based.